* general: finally activated translated messages for COBOL runtime
libcob:
* common.c, common.h: new function cob_common_init to provide C RTS and library setup, currently binding textdomain and for [_WIN32] handling unix-lf
* common.c, coblocal.h: activated translated messages for COBOL runtime
* common.c, screenio.c: fixed to use constant strings for static char * initialization, postpone msgid translation
* common.c (print_runtime_conf, get_config_val, var_print): adjusted output of cobcrun -runtime-conf (boolean print as yes/no now)

+++ cobc/typeck.c
@@ -8894,7 +8894,7 @@ validate_move (cb_tree src, cb_tree dst, const unsigned int is_value, int *move_
 					goto size_overflow;
 				}
 				/* for VALUE: additional check without trim */
-				if (is_value && l->size > fdst->size) {
+				if (is_value && l->size > (unsigned int)fdst->size) {
 					goto value_mismatch;
 				}
 			}
